refactor: replace new fall back image
