+++ src/components/Card/CardDialog.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { NFTDetail, NftType } from '../../models'
 import styled from 'styled-components'
 import { Limited } from '../Limited'
-import FallbackImg from '../../assets/img/card-fallback.png'
+import FallbackImg from '../../assets/svg/fallback.svg'
 import { getImagePreviewUrl } from '../../utils'
 import { LazyLoadImage } from '../Image'
 import { Like } from '../Like'
